requirements_verification.txt
torch>=1.5.1
torchvision>=0.6.1
ml_collections
apex
tqdm
tensorboard
scikit-learn>=0.24.0
numpy
scipy
Pillow
matplotlib
pandas 